Refactor NoSaveButton for typescript
